Add docs `Time::class`.
