+++ tests/Input/Time/AttributeTest.php
@@ -153,26 +153,6 @@ final class AttributeTest extends TestCase
         );
     }
 
-    public function testMax(): void
-    {
-        Assert::equalsWithoutLE(
-            <<<HTML
-            <input id="time-6582f2d099e8b" type="time" max="value">
-            HTML,
-            Time::widget()->id('time-6582f2d099e8b')->max('value')->render()
-        );
-    }
-
-    public function testMin(): void
-    {
-        Assert::equalsWithoutLE(
-            <<<HTML
-            <input id="time-6582f2d099e8b" type="time" min="value">
-            HTML,
-            Time::widget()->id('time-6582f2d099e8b')->min('value')->render()
-        );
-    }
-
     public function testName(): void
     {
         Assert::equalsWithoutLE(
@@ -190,26 +170,6 @@ final class AttributeTest extends TestCase
             <input id="time-6582f2d099e8b" type="time" readonly>
             HTML,
             Time::widget()->id('time-6582f2d099e8b')->readonly()->render()
-        );
-    }
-
-    public function testRender(): void
-    {
-        Assert::equalsWithoutLE(
-            <<<HTML
-            <input id="time-6582f2d099e8b" type="time">
-            HTML,
-            Time::widget()->id('time-6582f2d099e8b')->render()
-        );
-    }
-
-    public function testRequired(): void
-    {
-        Assert::equalsWithoutLE(
-            <<<HTML
-            <input id="time-6582f2d099e8b" type="time" required>
-            HTML,
-            Time::widget()->id('time-6582f2d099e8b')->required()->render()
         );
     }
